requirements.txt
Flask
pypdf
pillow
gunicorn
